Move Quest2 (now Quest) to quests.quest. Remove unused code.
