+++ src/data/quests/first_quest.yml
