+++ src/data/quests/sample_quest.yml
