+++ src/main.py
@@ -1,150 +1,56 @@
 import sys
-from typing import Union
 
-import networkx
 import pygame as pg
 
 import controller
 import images
 import settings
 import sounds
-from creatures import players
-from creatures.humanoids import Inventory, Status, HumanoidData
 from draw_utils import draw_text
-from quests.resolutions import Resolution
-from quests.scenes import Scene, make_scene
+from quests.quest import Quest
 
+lose_data = {'type': 'decision',
+             'prompt': 'You lose!',
+             'choices': [{'play again': 'root'}]}
 
-class Quest2(object):
-    """Handles transitions between different Controllers (scenes)"""
+win_data = lose_data.copy()
+win_data['prompt'] = 'you win!'
 
-    def __init__(self):
+rock_resolutions = [{'kill group': {'group label': 'quest',
+                                    'next scene': 'game over win'}},
+                    {'condition': {'condition data': {'dead': None},
+                                   'tested label': 'player',
+                                   'next scene': 'game over lose'}},
+                    {'enter zone': {'zone label': 'exit',
+                                    'entering label': 'player',
+                                    'next scene': 'game over win'}}]
 
-        self._player_data: HumanoidData = None
-        self._graph = networkx.MultiDiGraph()
+laser_resolutions = [{'kill group': {'group label': 'quest',
+                                     'next scene': 'rock scene'}},
+                     {'condition': {'condition data': {'dead': None},
+                                    'tested label': 'player',
+                                    'next scene': 'game over lose'}},
+                     {'enter zone': {'zone label': 'exit',
+                                     'entering label': 'player',
+                                     'next scene': 'rock scene'}}]
 
-        lose_data = {'type': 'decision',
-                     'prompt': 'You lose!',
-                     'choices': [{'play again': 'root'}]}
+laser_scene_data = {'type': 'dungeon',
+                    'map file': 'level1.tmx',
+                    'resolutions': laser_resolutions}
+rock_scene_data = {'type': 'dungeon',
+                   'map file': 'goto.tmx',
+                   'resolutions': rock_resolutions}
 
-        win_data = lose_data.copy()
-        win_data['prompt'] = 'you win!'
+start_scene_data = {'type': 'decision',
+                    'prompt': 'Lasers or rocks?',
+                    'choices': [{'lasers please': 'laser scene'},
+                                {'rocks!': 'rock scene'}]}
 
-        rock_resolutions = [{'kill group': {'group label': 'quest',
-                                            'next scene': 'game over win'}},
-                            {'condition': {'condition data': {'dead': None},
-                                           'tested label': 'player',
-                                           'next scene': 'game over lose'}},
-                            {'enter zone': {'zone label': 'exit',
-                                            'entering label': 'player',
-                                            'next scene': 'game over win'}}]
-
-        laser_resolutions = [{'kill group': {'group label': 'quest',
-                                             'next scene': 'rock scene'}},
-                             {'condition': {'condition data': {'dead': None},
-                                            'tested label': 'player',
-                                            'next scene': 'game over lose'}},
-                             {'enter zone': {'zone label': 'exit',
-                                             'entering label': 'player',
-                                             'next scene': 'rock scene'}}]
-
-        laser_scene_data = {'type': 'dungeon',
-                            'map file': 'level1.tmx',
-                            'resolutions': laser_resolutions}
-        rock_scene_data = {'type': 'dungeon',
-                           'map file': 'goto.tmx',
-                           'resolutions': rock_resolutions}
-
-        start_scene_data = {'type': 'decision',
-                            'prompt': 'Lasers or rocks?',
-                            'choices': [{'lasers please': 'laser scene'},
-                                        {'rocks!': 'rock scene'}]}
-
-        quest_data = {'root': start_scene_data,
-                      'laser scene': laser_scene_data,
-                      'rock scene': rock_scene_data,
-                      'game over win': win_data,
-                      'game over lose': lose_data}
-
-        g2 = networkx.MultiDiGraph()
-        for label, scene_data in quest_data.items():
-            g2.add_node(make_scene(scene_data), label=label)
-
-        scene_from_label = {info['label']: node
-                            for node, info in g2.nodes.data()}
-
-        for scene_label, scene in scene_from_label.items():
-            scene_data = quest_data[scene_label]
-
-            if scene_data['type'] == 'decision':
-                for index, choice in enumerate(scene_data['choices']):
-                    assert len(choice.values()) == 1
-                    next_scene_label = list(choice.values())[0]
-                    next_scene = scene_from_label[next_scene_label]
-                    g2.add_edge(scene, next_scene, key=index)
-
-            else:
-                assert scene_data['type'] == 'dungeon'
-                for index, resolution in enumerate(scene_data['resolutions']):
-                    assert len(resolution.values()) == 1
-                    res_data = list(resolution.values())[0]
-                    next_scene_label = res_data['next scene']
-                    next_scene = scene_from_label[next_scene_label]
-                    g2.add_edge(scene, next_scene, key=index)
-
-        root = scene_from_label['root']
-        self._graph = g2
-        self._root_scene = root
-        self._set_current_scene(root)
-
-    def _set_current_scene(self, scene: Scene) -> None:
-        self._current_scene = scene
-        ctrl, resolutions = self._current_scene.make_controller_and_resolutions()
-        self._current_ctrl = ctrl
-
-        if self._current_scene is self._root_scene:
-            self._player_data = HumanoidData(Status(players.PLAYER_HEALTH),
-                                             Inventory())
-        self._current_ctrl.set_player_data(self._player_data)
-
-        resols = self._resolution_to_next_scene_map(scene, resolutions)
-        self._resolutions_to_scenes = resols
-
-    def _resolution_to_next_scene_map(self, current_scene, resolutions):
-        """ The Scene object outputs resolutions in a specific order. We match
-        that order to the key assigned to each edge, which tells us what scene
-        each resolution points to."""
-
-        # Output of out_edges is a list of tuples of the form
-        # (source :Scene, sink : Scene, key : int)
-        next_scenes = [(edge[2], edge[1]) for edge in
-                       self._graph.out_edges(current_scene, keys=True)]
-        next_scenes = sorted(next_scenes, key=lambda x: x[0])
-        resols = {res: scene_tup[1] for res, scene_tup in
-                  zip(resolutions, next_scenes)}
-        return resols
-
-    def update_and_draw(self):
-        self._current_ctrl.update()
-        self._current_ctrl.draw()
-
-        resolution = self._resolved_resolution()
-        if resolution is not None:
-            next_scene = self._resolutions_to_scenes[resolution]
-            self._set_current_scene(next_scene)
-
-    def _resolved_resolution(self) -> Union[Resolution, None]:
-        resolved = [res for res in self._resolutions_to_scenes if
-                    res.is_resolved]
-
-        if len(resolved) not in (0, 1):
-            raise Warning('More than one resolved resolutions: {}. '
-                          'Choosing first in list.'.format(resolved))
-
-        if resolved:
-            return resolved[0]
-        else:
-            return None
+quest_data_dict = {'root': start_scene_data,
+                   'laser scene': laser_scene_data,
+                   'rock scene': rock_scene_data,
+                   'game over win': win_data,
+                   'game over lose': lose_data}
 
 
 class Game(object):
@@ -174,7 +80,7 @@ class Game(object):
 
     def new(self) -> None:
 
-        self.quest_graph = Quest2()
+        self.quest_graph = Quest(quest_data_dict)
 
         sounds.play(sounds.LEVEL_START)
